Add trait RuleUtilsTest
