+++ src/test/scala/com/microsoft/hyperspace/index/rules/JoinIndexRuleTest.scala
@@ -17,18 +17,17 @@
 package com.microsoft.hyperspace.index.rules
 
 import org.apache.hadoop.fs.Path
-import org.apache.spark.sql.{Row, SparkSession}
+import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.plans.{JoinType, SQLHelper}
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.execution.datasources._
-import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
-import org.apache.spark.sql.types.{IntegerType, StringType, StructType}
+import org.apache.spark.sql.types.{IntegerType, StringType}
 
 import com.microsoft.hyperspace.index._
 import com.microsoft.hyperspace.util.FileUtils
 
-class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
+class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper with RuleTestUtils {
   override val systemPath = new Path("src/test/resources/joinIndexRuleTest")
 
   val t1c1 = AttributeReference("t1c1", IntegerType)()
@@ -40,8 +39,8 @@ class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
   val t2c3 = AttributeReference("t2c3", IntegerType)()
   val t2c4 = AttributeReference("t2c4", StringType)()
 
-  val t1Schema = RuleTestHelper.schemaFromAttributes(t1c1, t1c2, t1c3, t1c4)
-  val t2Schema = RuleTestHelper.schemaFromAttributes(t2c1, t2c2, t2c3, t2c4)
+  val t1Schema = schemaFromAttributes(t1c1, t1c2, t1c3, t1c4)
+  val t2Schema = schemaFromAttributes(t2c1, t2c2, t2c3, t2c4)
 
   var t1Relation: HadoopFsRelation = _
   var t2Relation: HadoopFsRelation = _
@@ -72,8 +71,8 @@ class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
     val t2Location =
       new InMemoryFileIndex(spark, Seq(new Path("t2")), Map.empty, Some(t2Schema), NoopCache)
 
-    t1Relation = RuleTestHelper.baseRelation(t1Location, t1Schema, spark)
-    t2Relation = RuleTestHelper.baseRelation(t2Location, t2Schema, spark)
+    t1Relation = baseRelation(t1Location, t1Schema, spark)
+    t2Relation = baseRelation(t2Location, t2Schema, spark)
 
     t1ScanNode = LogicalRelation(t1Relation, Seq(t1c1, t1c2, t1c3, t1c4), None, false)
     t2ScanNode = LogicalRelation(t2Relation, Seq(t2c1, t2c2, t2c3, t2c4), None, false)
@@ -91,11 +90,11 @@ class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
     //  +- Filter isnotnull(t2c1#4)
     //   +- Relation[t2c1#4,t2c2#5,t2c3#6,t2c4#7] parquet
 
-    RuleTestHelper.createIndex(systemPath, spark, "t1i1", Seq(t1c1), Seq(t1c3), t1ProjectNode)
-    RuleTestHelper.createIndex(systemPath, spark, "t1i2", Seq(t1c1, t1c2), Seq(t1c3), t1ProjectNode)
-    RuleTestHelper.createIndex(systemPath, spark, "t1i3", Seq(t1c2), Seq(t1c3), t1ProjectNode)
-    RuleTestHelper.createIndex(systemPath, spark, "t2i1", Seq(t2c1), Seq(t2c3), t2ProjectNode)
-    RuleTestHelper.createIndex(systemPath, spark, "t2i2", Seq(t2c1, t2c2), Seq(t2c3), t2ProjectNode)
+    createIndex(systemPath, spark, "t1i1", Seq(t1c1), Seq(t1c3), t1ProjectNode)
+    createIndex(systemPath, spark, "t1i2", Seq(t1c1, t1c2), Seq(t1c3), t1ProjectNode)
+    createIndex(systemPath, spark, "t1i3", Seq(t1c2), Seq(t1c3), t1ProjectNode)
+    createIndex(systemPath, spark, "t2i1", Seq(t2c1), Seq(t2c3), t2ProjectNode)
+    createIndex(systemPath, spark, "t2i2", Seq(t2c1, t2c2), Seq(t2c3), t2ProjectNode)
   }
 
   before {
@@ -196,10 +195,11 @@ class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
 
     {
       // Test: should update plan if index exists to cover all implicit columns
-      val t1TestIndex = RuleTestHelper
-        .createIndex(systemPath, spark, "t1Idx", Seq(t1c1), Seq(t1c2, t1c3, t1c4), t1FilterNode)
-      val t2TestIndex = RuleTestHelper
-        .createIndex(systemPath, spark, "t2Idx", Seq(t2c1), Seq(t2c2, t2c3, t2c4), t2FilterNode)
+      val t1TestIndex =
+        createIndex(systemPath, spark, "t1Idx", Seq(t1c1), Seq(t1c2, t1c3, t1c4), t1FilterNode)
+
+      val t2TestIndex =
+        createIndex(systemPath, spark, "t2Idx", Seq(t2c1), Seq(t2c2, t2c3, t2c4), t2FilterNode)
 
       // clear cache so the new indexes gets added to it
       clearCache()
@@ -409,9 +409,8 @@ class JoinIndexRuleTest extends HyperspaceSuite with SQLHelper {
       .flatten
   }
 
-
   private def getIndexDataFilesPath(indexName: String): Path = {
-    RuleTestHelper.getIndexDataFilesPath(indexName, systemPath)
+    getIndexDataFilesPathImpl(indexName, systemPath)
   }
 
   private def getIndexRootPath(indexName: String): Path = {
